bringing changes in line

+++ scripts/rosmap_cit_replication.py
@@ -38,7 +38,7 @@ def cit_on_qtl_set(df, gene, coord_df, methyl, acetyl, express, opts):
     # reduce genotype
     latent_genotype = dm.reduce_genotype(cur_genotype, opts['lv_method'], opts['num_latent'], opts['vae_depth'])
     if type(latent_genotype) != numpy.ndarray:
-        latent_genotype = latent_genotype.numpy()
+        latent_genotype = latent_genotype.numpy().astype(np.float64)
     # get probes and peaks
     cur_exp = cur_expression[:, e_ids == gene]
